backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.6
pandas==2.2.0
openpyxl==3.1.2
gspread==6.0.0
oauth2client==4.1.3
requests==2.31.0
email-validator
jinja2==3.1.3
itsdangerous==2.1.2
